requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.3
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pymongo==4.16.0
python-dotenv==1.2.1
Werkzeug==3.1.6
openpyxl>=3.1.2
APScheduler>=3.10.4
bcrypt>=4.1.2
email-validator>=2.1.1
